+++ src/VEF/Turbulence/Op_Dift_VEF_Face_Q1.h
@@ -58,7 +58,7 @@ public:
 
   inline void remplir_nu(DoubleTab& ) const
   {
-    Cerr<<__FILE__<<":"<<__LINE__<<finl;
+    Cerr<<__FILE__<<":"<<(int)__LINE__<<finl;
     exit();
   } ;
